% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hydrus_exe.R
\name{check_hydrus_exe}
\alias{check_hydrus_exe}
\title{Check HYDRUS Executable and download if needed}
\usage{
check_hydrus_exe(
  dir = file.path(system.file(package = "kwb.hydrus1d"), "extdata/hydrus1d"),
  skip_preinstalled = FALSE
)
}
\arguments{
\item{dir}{target directory (default: file.path(system.file(package =
"kwb.hydrus1d"), "extdata/hydrus1d"), in case compliant HYDRUS1D version is
not available, executable will be downloaded from
\url{https://github.com/mrustl/hydrus1d/archive/refs/tags/v4.17.0140.zip}}

\item{skip_preinstalled}{if TRUE checking of preinstalled HYDRUS-1D will be
skipped and executable will be downloaded, otherwise not (the default)}
}
\value{
path to preinstalled HYDRUS-1D executable (if compliant and parameter
skip_preinstalled == FALSE), otherwise path to downloaded HYDRUS-1D
}
\description{
Check HYDRUS Executable and download if needed
}
\examples{
check_hydrus_exe(skip_preinstalled = FALSE)
check_hydrus_exe(skip_preinstalled = TRUE)
}
