% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_atmosphere_to_string.R
\name{convert_atmosphere_to_string}
\alias{convert_atmosphere_to_string}
\title{Convert Atmosphere to String}
\usage{
convert_atmosphere_to_string(atm, round_digits = 2L, remove_scientific = TRUE)
}
\arguments{
\item{atm}{atm as retrieved by \code{prepare_atmosphere_data}}

\item{round_digits}{digits used for rounding values (default: 2) of all columns
besides "tAtm"}

\item{remove_scientific}{if TRUE scientific notation of numbers is removed,
otherwise not (default: TRUE)}
}
\value{
properly padded string of atmosphere variables for Hydrus1D input file
}
\description{
Convert Atmosphere to String
}
