% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults-atmosphere.R
\name{defaults_atmosphere}
\alias{defaults_atmosphere}
\title{Defaults for Atmosphere}
\usage{
defaults_atmosphere(
  Prec = 0L,
  rSoil = 0L,
  rRoot = 0L,
  hCritA = 100000L,
  rB = 0L,
  hB = 0L,
  ht = 0L,
  tTop = 0L,
  tBot = 0L,
  Ampl = 0L,
  cTop = 0L,
  cBot = 0L,
  RootDepth = 0L
)
}
\arguments{
\item{Prec}{Precipitation (default: 0)}

\item{rSoil}{Evaporation (default: 0)}

\item{rRoot}{Transpiration (default: 0)}

\item{hCritA}{(default: 100000)}

\item{rB}{rB (default: 0)}

\item{hB}{hB (default: 0)}

\item{ht}{ht (default: 0)}

\item{tTop}{tTop (default: 0)}

\item{tBot}{tBot (default: 0)}

\item{Ampl}{Ampl (default: 0)}

\item{cTop}{concentration of solute 1 at top (default: 0)}

\item{cBot}{concentration of solute 1 at bottom (default: 0)}

\item{RootDepth}{root depth (default: 0)}
}
\value{
tibble with defaults for atmospheric parameters
}
\description{
Defaults for Atmosphere
}
\examples{
defaults_atmosphere()
}
