% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_atmosphere_input.R
\name{get_atmosphere_headers}
\alias{get_atmosphere_headers}
\title{Get Atmosphere Headers}
\usage{
get_atmosphere_headers()
}
\value{
vector with atmosphere headers for "ATMOSPH.in"
}
\description{
Get Atmosphere Headers
}
\examples{
get_atmosphere_headers()
}
