% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_number_of_endlines.R
\name{get_number_of_endlines}
\alias{get_number_of_endlines}
\title{Helper function: get number of "endlines"}
\usage{
get_number_of_endlines(content)
}
\arguments{
\item{content}{content of text file}
}
\value{
returns number of lines starting with "end"
}
\description{
Helper function: get number of "endlines"
}
