% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_units_list.R
\name{get_units_list}
\alias{get_units_list}
\title{Helper function: get units list from meta general}
\usage{
get_units_list(meta_general)
}
\arguments{
\item{meta_general}{as retrieved by \code{read_meta_general}}
}
\value{
list with elements unit_L, unit_T and unit_M
}
\description{
Helper function: get units list from meta general
}
\examples{
path_tlevel <- system.file("extdata/model/test/T_LEVEL.out", package = "kwb.hydrus1d")
content <- readLines(path_tlevel)
content_general <- content[3:5]
cat(content_general)
meta_general <- read_meta_general(content_general)
get_units_list(meta_general)

}
