% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_atmosphere_input.R
\name{prepare_atmosphere_input}
\alias{prepare_atmosphere_input}
\title{Prepare Atmosphere Input}
\usage{
prepare_atmosphere_input(inputs, defaults = defaults_atmosphere())
}
\arguments{
\item{inputs}{tibble or data.frame with user defined inputs for parameters
defined in \code{get_atmosphere_headers}}

\item{defaults}{defaults for undefined parameters. Default:
\code{\link{defaults_atmosphere}}}
}
\value{
tibble with atmosphere values
}
\description{
Prepare Atmosphere Input
}
\examples{
inputs <- tibble::tibble(tAtm = 1:10, Prec = 10, rSoil = 0.4)
prepare_atmosphere_input(inputs)
}
