% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevel.R
\name{read_alevel}
\alias{read_alevel}
\title{Read A_LEVEL.out}
\usage{
read_alevel(path)
}
\arguments{
\item{path}{full path to A_LEVEL.out file}
}
\value{
imports A_LEVEL out with tidy column names
}
\description{
Read A_LEVEL.out
}
