% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_atmosph.R
\name{read_atmosph}
\alias{read_atmosph}
\title{Read ATMOSPH.in}
\usage{
read_atmosph(
  path = system.file("extdata/model/test/ATMOSPH.in", package = "kwb.hydrus1d")
)
}
\arguments{
\item{path}{path to ATMOSPH.in (default: system.file("extdata/model/test/ATMOSPH.in",
package = "kwb.hydrus1d"))}
}
\value{
list with atmospheric config parameter sublist and time series sublist
}
\description{
Read ATMOSPH.in
}
