% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_balance.R
\name{read_balance}
\alias{read_balance}
\title{Read BALANCE.out}
\usage{
read_balance(
  path = system.file("extdata/model/test/BALANCE.out", package = "kwb.hydrus1d")
)
}
\arguments{
\item{path}{full path to BALANCE.out file (default: system.file("extdata/model/test/BALANCE.out",
package = "kwb.hydrus1d"))}
}
\value{
tibble with "balance" time series
run
}
\description{
Reads BALANCE.out output
}
