% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_hydrus1d_dat.R
\name{read_hydrus1d}
\alias{read_hydrus1d}
\title{Read HYDRUS1D.dat file}
\usage{
read_hydrus1d(
  path = system.file("extdata/model/test/HYDRUS1D.dat", package = "kwb.hydrus1d")
)
}
\arguments{
\item{path}{path to HYDRUS1D.dat file (default: system.file("extdata/model/test/HYDRUS1D.dat",
package = "kwb.hydrus1d"))}
}
\value{
list with parameters defined in
}
\description{
Read HYDRUS1D.dat file
}
