% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meta_general.R
\name{read_meta_general}
\alias{read_meta_general}
\title{Helper Function: Read Meta General}
\usage{
read_meta_general(content_general)
}
\arguments{
\item{content_general}{lines with content general (see example)}
}
\value{
tibble with columns description, modelstart_datetime and unit L,T,M
}
\description{
Helper Function: Read Meta General
}
\examples{
path_tlevel <- system.file("extdata/model/test/T_LEVEL.out", package = "kwb.hydrus1d")
content <- readLines(path_tlevel)
content_general <- content[3:5]
cat(content_general)
read_meta_general(content_general)
}
