% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obsnode.R
\name{read_obsnode}
\alias{read_obsnode}
\title{Read Obs_Node.out}
\usage{
read_obsnode(path, to_longer = TRUE)
}
\arguments{
\item{path}{path to Obs_Node.out}

\item{to_longer}{convert table to longer format (default: TRUE)}
}
\description{
Read Obs_Node.out
}
