% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_profile.R
\name{read_profile}
\alias{read_profile}
\title{Read PROFILE.out}
\usage{
read_profile(path)
}
\arguments{
\item{path}{path to PROFILE.out}
}
\value{
tibble with PROFILE.out data
}
\description{
Read PROFILE.out
}
