% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_runinf.R
\name{read_runinf}
\alias{read_runinf}
\title{Read Run_Inf.out}
\usage{
read_runinf(path)
}
\arguments{
\item{path}{full path to Run_Inf.out file}
}
\value{
imports Run_Inf.out with tidy column names and saves metainformation
in attribute 'meta'
\describe{
\item{tlevel}{Time-level (current time-step number) [-].}
\item{time}{Time, t, at current time-level \link{T}.}
}
}
\description{
Contains time and iteration information
}
\examples{
path_runinf <- system.file("extdata/model/test/Run_Inf.out", package = "kwb.hydrus1d")
runinf <- read_runinf(path = path_runinf)
runinf
}
\references{
\url{https://www.pc-progress.com/Downloads/Pgm_Hydrus1D/HYDRUS1D-4.17.pdf#page=272}
}
