% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_solute.R
\name{read_solute}
\alias{read_solute}
\title{Read solute1.out}
\usage{
read_solute(
  path = system.file("extdata/model/test/solute1.out", package = "kwb.hydrus1d")
)
}
\arguments{
\item{path}{full path to solute_id.out file (default: system.file("extdata/model/test/solute1.out",
package = "kwb.hydrus1d"))}
}
\value{
imports solute_id.out with tidy column names and saves metainformation
in attributes 'meta_general' and 'meta_units'
\describe{
\item{time}{Time, t, at current time-level [T]}
\item{cv_top}{Actual solute flux across the soil surface [ML-2 T -1] (inflow/outflow: +/-)}
\item{cv_bot}{Actual solute flux across the bottom of the soil profile [ML-2 T -1]
(inflow/outflow: +/-)}
\item{sum_cv_top}{Cumulative solute flux across the soil surface [ML-2]
(inflow/outflow: +/-)}
\item{sum_cv_bot}{Cumulative solute flux across the bottom of the soil
profile [ML-2] (inflow/outflow: +/-)}
\item{sum_cv_ch0}{Cumulative amount of solute added to the flow region by
zero-order reactions [ML-2] (negative when removed from the system)}
\item{sum_cv_ch1}{Cumulative amount of solute removed from the flow region
by first-order reactions [ML-2] (negative removed from the system).}
\item{c_top}{Solute concentration at the soil surface [ML-3]}
\item{c_root}{Mean solute concentration of the root zone [ML-3]}
\item{c_bot}{Solute concentration at the bottom of the soil profile [ML-3]}
\item{cv_root}{Actual root solute uptake in the root zone [ML-2 T -1]
(positive when removed from the system)}
\item{sum_cv_root}{Cumulative amount of solute removed from the flow region
by root water uptake S [ML-2] (positive when removed from the system)}
\item{sum_cv_n_eql}{Cumulative mass transfer to either kinetic adsorption
sites (type-2 adsorption sites), or to the immobile liquid region [ML-2]
(inflow/outflow: +/-).}
\item{t_level}{Time-level (current time-step number) [-]}
\item{c_gwl}{Average concentration in the saturated zone [ML-3] (in the
groundwater)}
\item{c_run_off}{Solute flux in the runoff ([ML-3] * [LT-1]) [ML-2T-1]}
\item{sum_c_run_off}{Cumulative solute flux in the runoff ([ML-3] * [LT-1]* [T])
[ML-2]}
\item{cv_i}{Solute flux at the first through third observation node
([ML-3] * [LT-1]) [ML-2T-1]}
\item{sum_cv_i}{cumulative solute flux at the first through third observation
node ([ML-3] * [LT-1] * [T]) [ML-2]. Total solute fluxes are reported
when only one solute is simulated. Only convective fluxes (for the first
solute) are reported when 2 or more solutes are simulated}
}
}
\description{
Reads Solute output
}
\examples{
path_solute <- system.file("extdata/model/test/solute1.out", package = "kwb.hydrus1d")
solute <- read_solute(path = path_solute)
solute
}
\references{
\url{https://www.pc-progress.com/Downloads/Pgm_Hydrus1D/HYDRUS1D-4.17.pdf#page=273}
}
