% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tlevel.R
\name{read_tlevel}
\alias{read_tlevel}
\title{Read T_LEVEL.out}
\usage{
read_tlevel(path)
}
\arguments{
\item{path}{full path to T_LEVEL.out file}
}
\value{
imports T_LEVEL out with tidy column names and saves metainformation
in attribute 'meta'
}
\description{
Read T_LEVEL.out
}
\examples{
path_tlevel <- system.file("extdata/model/test/T_LEVEL.out", package = "kwb.hydrus1d")
tlevel <- read_tlevel(path = path_tlevel)
tlevel
}
