% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metadata.R
\name{set_metadata}
\alias{set_metadata}
\title{Set Metadata Attributes}
\usage{
set_metadata(x, meta_general, meta_units)
}
\arguments{
\item{x}{R object}

\item{meta_general}{general metadata}

\item{meta_units}{information about units}
}
\value{
\code{x} with attributes "meta_general" and "meta_units" being set
(may be \code{NULL})
}
\description{
Set Metadata Attributes
}
