% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tlevel.R
\name{tlevel_get_meta}
\alias{tlevel_get_meta}
\title{Helper Function: Get Metadata of 'T_LEVEL.out'}
\usage{
tlevel_get_meta(tlevel)
}
\arguments{
\item{tlevel}{as retrieved by \code{read_tlevel}}
}
\value{
returns metainformation list with sublists "general" and "units"  of
imported T_LEVEL.out file
}
\description{
Helper Function: Get Metadata of 'T_LEVEL.out'
}
\examples{
path_tlevel <- system.file("extdata/model/test/T_LEVEL.out", package = "kwb.hydrus1d")
tlevel <- read_tlevel(path = path_tlevel)
tlevel_get_meta(tlevel)
}
