% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_atmosphere.R
\name{write_atmosphere}
\alias{write_atmosphere}
\title{Write "ATMOSPH.IN" input file}
\usage{
write_atmosphere(
  atm,
  MaxAL = nrow(atm),
  DailyVar = FALSE,
  SinusVar = FALSE,
  lLai = FALSE,
  lBCCycles = FALSE,
  lInterc = FALSE,
  hCritS = 0,
  round_digits = 2,
  remove_scientific = TRUE
)
}
\arguments{
\item{atm}{tibble of input data as defined in \code{prepare_atmospherice_input}}

\item{MaxAL}{Number of meteorological records (default: nrow(atm))}

\item{DailyVar}{TRUE if HYDRUS-1D is to generate daily variations in evaporation
and transpiration (see section 2.7.2.)., otherwise: FALSE (default: FALSE)}

\item{SinusVar}{TRUE if HYDRUS-1D is to generate sinusoidal variations in
precipitation (see section 2.7.2.), otherwise: FALSE(default: FALSE)}

\item{lLai}{Logical variable indicating that potential evapotranspiration is
to be divided into potential evaporation and potential transpiration using
eq. (2.75). (default: FALSE)}

\item{lBCCycles}{TRUE if a set of boundary conditions is to be repeated
multiple times, otherwise FALSE(default: FALSE)}

\item{lInterc}{TRUE if interception is considered using eq. (2.78), otherwise
FALSE (default: FALSE)}

\item{hCritS}{Maximum allowed pressure head at the soil surface (L). (default: 0)}

\item{round_digits}{digits used for rounding values (default: 2) of all columns
besides "tAtm"}

\item{remove_scientific}{if TRUE scientific notation of numbers is removed,
otherwise not (default: TRUE)}
}
\value{
Creates ATMOSPH.IN input textfile
}
\description{
Write "ATMOSPH.IN" input file
}
\examples{
inputs <- tibble::tibble(tAtm = 1:10, Prec = 10, rSoil = 0.4)
atm <- prepare_atmosphere_input(inputs)
atm
atm_string <- write_atmosphere(atm = atm, MaxAL = nrow(atm))
cat(atm_string)
}
