% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_profile.R
\name{write_formatted_materials}
\alias{write_formatted_materials}
\title{Helper function: write formatted soil materials}
\usage{
write_formatted_materials(df)
}
\arguments{
\item{df}{df with soil materials as retrieved by \code{read_profile} and
sublist "mat_props"}
}
\value{
formatted soil materials
}
\description{
Helper function: write formatted soil materials
}
\keyword{internal}
