% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_profile.R
\name{write_formatted_profile}
\alias{write_formatted_profile}
\title{Helper function: write formatted soil profile}
\usage{
write_formatted_profile(df)
}
\arguments{
\item{df}{df with soil profile as retrieved by \code{read_profile} and sublist
profile}
}
\value{
formatted soil materials
}
\description{
Helper function: write formatted soil profile
}
\keyword{internal}
