% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_hydrus1d_dat.R
\name{write_hydrus1d}
\alias{write_hydrus1d}
\title{Write HYDRUS1D.dat}
\usage{
write_hydrus1d(
  hydrus1d_list,
  path = system.file("extdata/model/test/HYDRUS1D.dat", package = "kwb.hydrus1d")
)
}
\arguments{
\item{hydrus1d_list}{as retrieved by \code{read_hydrus1d}}

\item{path}{path to HYDRUS1D.dat for export (default: system.file("extdata/model/test/HYDRUS1D.dat",
package = "kwb.hydrus1d")}
}
\value{
write HYDRUS1D.dat
}
\description{
Write HYDRUS1D.dat
}
