% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_profile.R
\name{write_profile}
\alias{write_profile}
\title{Write PROFILE.dat}
\usage{
write_profile(profile, path)
}
\arguments{
\item{profile}{profile in structure as imported with \code{read_profile}}

\item{path}{path to export PROFILE.dat}
}
\value{
writes PROFILE.dat to user specified path
}
\description{
Write PROFILE.dat
}
