% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_selector.R
\name{write_selector}
\alias{write_selector}
\title{Write SELECTOR.in}
\usage{
write_selector(selector, path)
}
\arguments{
\item{selector}{list with imported SELECTOR.in \code{read_selector}}

\item{path}{path to SELECTOR.in for export}
}
\description{
Write SELECTOR.in
}
