% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_atmosphere_to_string.R
\name{convert_atmosphere_to_string}
\alias{convert_atmosphere_to_string}
\title{Convert Atmosphere to String}
\usage{
convert_atmosphere_to_string(atm)
}
\arguments{
\item{atm}{atm as retrieved by \code{prepare_atmosphere_data}}
}
\value{
tibble with strings padded properly for Hydrus1D
}
\description{
Convert Atmosphere to String
}
