% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_meta.R
\name{get_output_meta}
\alias{get_output_meta}
\title{Helper Function: Get Metadata of Outputs ('A_LEVEL.out', 'T_LEVEL.out')}
\usage{
get_output_meta(output)
}
\arguments{
\item{output}{imported output as retrieved by \code{read_alevel} or \code{read_tlevel}}
}
\value{
returns metainformation list with sublists "general" and "units"  of
imported 'A_LEVEL.out' or 'T_LEVEL.out' file
}
\description{
Helper Function: Get Metadata of Outputs ('A_LEVEL.out', 'T_LEVEL.out')
}
\examples{
path_alevel <- system.file("extdata/model/test/A_LEVEL.out", package = "kwb.hydrus1d")
alevel <- read_tlevel(path = path_tlevel)
get_output_meta(alevel)
path_tlevel <- system.file("extdata/model/test/T_LEVEL.out", package = "kwb.hydrus1d")
tlevel <- read_tlevel(path = path_tlevel)
get_output_meta(level)
}
