% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alevel.R
\name{read_alevel}
\alias{read_alevel}
\title{Read A_LEVEL.out}
\usage{
read_alevel(path)
}
\arguments{
\item{path}{full path to A_LEVEL.out file}
}
\value{
imports A_LEVEL out with tidy column names
}
\description{
A-level information is printed each time a time-dependent
boundary condition is specified. The information is directed to output file
A_LEVEL.OUT (Table 13.6).
}
\references{
\url{https://www.pc-progress.com/Downloads/Pgm_Hydrus1D/HYDRUS1D-4.17.pdf#page=277}
}
