% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_profile.R
\name{read_profile}
\alias{read_profile}
\title{Read PROFILE.DAT}
\usage{
read_profile(path)
}
\arguments{
\item{path}{path to PROFILE.DAT}
}
\value{
list with sublists with material properties, profile and observations
data of PROFILE.DAT data
}
\description{
Read PROFILE.DAT
}
