% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_runinf.R
\name{read_runinf}
\alias{read_runinf}
\title{Read Run_Inf.out}
\usage{
read_runinf(path, dbg = FALSE)
}
\arguments{
\item{path}{full path to Run_Inf.out file}

\item{dbg}{show debug messages (default: FALSE)}
}
\value{
imports Run_Inf.out with tidy column names and saves metainformation
in attributes 'meta_general' and 'meta_units'
\describe{
\item{tlevel}{Time-level (current time-step number) [-]}
\item{time}{Time, t, at current time-level [T]}
\item{dt}{Time step, delta t [T]}
\item{itr_w}{Number of iterations necessary for solution of the water flow
equation [-]}
\item{itr_c}{Number of iterations necessary for solution of the solute
transport equation [-]}
\item{it_cum}{Cumulative number of iterations [-]}
\item{kod_t}{Code for the boundary condition at the soil surface}
\item{kod_b}{Code for the boundary condition at the bottom of the soil
profile}
\item{converg}{Information whether or not the numerical convergence was
achieved at the current time-level}
\item{peclet}{Maximum local Peclet number [-]}
\item{courant}{Maximum local Courant number [-]}
}
}
\description{
Contains time and iteration information
}
\examples{
path_runinf <- system.file("extdata/model/test/Run_Inf.out", package = "kwb.hydrus1d")
runinf <- read_runinf(path = path_runinf)
runinf
}
\references{
\url{https://www.pc-progress.com/Downloads/Pgm_Hydrus1D/HYDRUS1D-4.17.pdf#page=272}
}
